Multiple updates to upgrade software versions and improve security
- Update all software versions
- Add Cloud Build config file
- Upgrade BigQuery Legacy SQL to Standard SQL
- Use cluster default credentials instead of private service account key
