+++ app/cloudbuild.yaml
@@ -1,0 +1,5 @@
+steps:
+  - name: 'gcr.io/cloud-builders/docker'
+    args: [ 'build', '-t', 'us-central1-docker.pkg.dev/$PROJECT_ID/dashboard-demo/dashboard-demo', './app' ]
+images:
+  - 'us-central1-docker.pkg.dev/$PROJECT_ID/dashboard-demo/dashboard-demo'
