+++ app/dashboard/modules/__init__.py
@@ -1,4 +1,4 @@
-# Copyright Google Inc. 2017
+# Copyright Google Inc. 2021
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -8,24 +8,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-
-class BaseModule:
-
-    def __init__(self):
-        self.id = self.__module__
-
-    def fetch_data(self, state):
-        raise NotImplementedError
-
-    def make_plot(self, dataframe):
-        raise NotImplementedError
-
-    def update_plot(self, dataframe):
-        raise NotImplementedError
-
-    def busy(self):
-        raise NotImplementedError
-
-    def unbusy(self):
-        raise NotImplementedError
