+++ app/dashboard/modules/temperature.py
@@ -1,4 +1,4 @@
-# Copyright Google Inc. 2017
+# Copyright Google Inc. 2021
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -11,43 +11,45 @@
 
 
 import pandas as pd
-
-from bokeh.models import ColumnDataSource, HoverTool, Paragraph, DataRange1d
-from bokeh.plotting import figure
+from bokeh.models import ColumnDataSource, HoverTool, DataRange1d, Paragraph
 from bokeh.layouts import column
+from bokeh.palettes import Blues4
+from bokeh.plotting import figure
 
 from modules.base import BaseModule
-from utils import run_query
 from states import NAMES_TO_CODES
+from utils import run_query
 
 
 QUERY = """
     SELECT
-      YEAR(date) as year,
-      MONTH(date) as month,
-      DAY(date) as day,
-      AVG(prcp) AS prcp
-    FROM (
-      SELECT
-        STRING(date) AS date,
-        IF (element = 'PRCP', value/10, NULL) AS prcp
-      FROM
-        [bigquery-public-data:ghcn_d.ghcnd_%(year)s] AS weather
-      JOIN
-        [bigquery-public-data:ghcn_d.ghcnd_stations] as stations
-      ON
-        weather.id = stations.id
-      WHERE
-        stations.state = '%(state)s'
-    )
+      year,
+      mo as month,
+      da as day,
+      MAX(max) as max_temp,
+      MIN(min) as min_temp,
+      AVG(temp) as avg_temp,
+      state
+    FROM
+      `bigquery-public-data.noaa_gsod.gsod%(year)s` a
+    JOIN
+      `bigquery-public-data.noaa_gsod.stations` b
+    ON
+      a.stn=b.usaf
+      AND a.wban=b.wban
+    WHERE
+      state IS NOT NULL
+      AND max < 1000
+      AND country = 'US'
+      AND state = '%(state)s'
     GROUP BY
-      year, month, day
+      year, month, day, state
     ORDER BY
-      year, month, day
+      year, month, day, state
 """
 
 YEAR = 2016
-TITLE = 'Precipitation (mm) in %s:' % YEAR
+TITLE = "Temperatures (F) in %s:" % YEAR
 
 
 class Module(BaseModule):
@@ -59,25 +61,31 @@ class Module(BaseModule):
         self.title = None
 
     def fetch_data(self, state):
-        dataframe = run_query(
-            QUERY % {'state': NAMES_TO_CODES[state], 'year': YEAR},
-            cache_key=('precipitation-%s' % NAMES_TO_CODES[state]))
+        query = QUERY % {'state': NAMES_TO_CODES[state], 'year': YEAR}
+        dataframe = run_query(query, cache_key='temperature-%s' % NAMES_TO_CODES[state])
         dataframe['date'] = pd.to_datetime(dataframe[['year', 'month', 'day']])
         dataframe['date_readable'] = dataframe['date'].apply(lambda x: x.strftime("%Y-%m-%d"))
+        dataframe['left'] = dataframe.date - pd.DateOffset(days=0.5)
+        dataframe['right'] = dataframe.date + pd.DateOffset(days=0.5)
+        dataframe = dataframe.set_index(['date'])
+        dataframe.sort_index(inplace=True)
         return dataframe
 
     def make_plot(self, dataframe):
         self.source = ColumnDataSource(data=dataframe)
         self.plot = figure(
-            x_axis_type="datetime", plot_width=400, plot_height=300,
+            x_axis_type="datetime", plot_width=600, plot_height=300,
             tools='', toolbar_location=None)
-
-        vbar = self.plot.vbar(
-            x='date', top='prcp', width=1, color='#fdae61', source=self.source)
+        self.plot.quad(
+            top='max_temp', bottom='min_temp', left='left', right='right',
+            color=Blues4[2], source=self.source, legend='Magnitude')
+        line = self.plot.line(
+            x='date', y='avg_temp', line_width=3, color=Blues4[1],
+            source=self.source, legend='Average')
         hover_tool = HoverTool(tooltips=[
             ('Value', '$y'),
             ('Date', '@date_readable'),
-        ], renderers=[vbar])
+        ], renderers=[line])
         self.plot.tools.append(hover_tool)
 
         self.plot.xaxis.axis_label = None
