+++ app/requirements.txt
@@ -1,0 +1,8 @@
+bokeh==2.4.1
+tornado==6.1
+pandas==1.3.4
+google-cloud-bigquery==2.30.1
+google-api-python-client==1.6.2
+scipy==1.7.2
+pymemcache==3.5.0
+pyarrow==6.0.0  # Required for BigQuery's to_dataframe() method
