--- dashboard/requirements.txt
@@ -1,7 +1,0 @@
-bokeh==0.12.5
-Tornado==4.4.2
-pandas==0.23.4
-pandas-gbq==0.6.0
-google-api-python-client==1.6.2
-scipy==0.19.0
-pymemcache==1.4.3
