+++ kubernetes/memcached.yaml
@@ -1,4 +1,4 @@
-# Copyright Google Inc. 2017
+# Copyright Google Inc. 2021
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -26,7 +26,7 @@ spec:
     spec:
       containers:
       - name: memcached
-        image: memcached:1.4.37
+        image: memcached:1.6.12
         ports:
         - containerPort: 11211
